+++ src/main/java/uk/gov/hmcts/reform/unspec/service/docmosis/aos/AcknowledgementOfServiceGenerator.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.unspec.model.CaseData;
 import uk.gov.hmcts.reform.unspec.model.Party;
-import uk.gov.hmcts.reform.unspec.model.SolicitorReferences;
 import uk.gov.hmcts.reform.unspec.model.docmosis.DocmosisDocument;
 import uk.gov.hmcts.reform.unspec.model.docmosis.aos.AcknowledgementOfServiceForm;
 import uk.gov.hmcts.reform.unspec.model.docmosis.sealedclaim.Respondent;
@@ -14,9 +13,8 @@ import uk.gov.hmcts.reform.unspec.model.documents.PDF;
 import uk.gov.hmcts.reform.unspec.service.docmosis.DocumentGeneratorService;
 import uk.gov.hmcts.reform.unspec.service.docmosis.TemplateDataGenerator;
 import uk.gov.hmcts.reform.unspec.service.documentmanagement.DocumentManagementService;
-import uk.gov.hmcts.reform.unspec.utils.CaseNameUtils;
+import uk.gov.hmcts.reform.unspec.utils.DocmosisTemplateDataUtils;
 
-import static java.util.Optional.ofNullable;
 import static uk.gov.hmcts.reform.unspec.service.docmosis.DocmosisTemplates.N9;
 
 @Service
@@ -43,26 +41,12 @@ public class AcknowledgementOfServiceGenerator implements TemplateDataGenerator<
     @Override
     public AcknowledgementOfServiceForm getTemplateData(CaseData caseData) {
         return AcknowledgementOfServiceForm.builder()
-            .caseName(CaseNameUtils.toCaseName.apply(caseData))
+            .caseName(DocmosisTemplateDataUtils.toCaseName.apply(caseData))
             .referenceNumber(caseData.getLegacyCaseReference())
-            .solicitorReferences(prepareSolicitorReferences(caseData.getSolicitorReferences()))
+            .solicitorReferences(DocmosisTemplateDataUtils.fetchSolicitorReferences(caseData.getSolicitorReferences()))
             .claimIssuedDate(caseData.getClaimIssuedDate())
             .responseDeadline(caseData.getRespondentSolicitor1ResponseDeadline().toLocalDate())
             .respondent(prepareRespondent(caseData.getRespondent1()))
-            .build();
-    }
-
-    public SolicitorReferences prepareSolicitorReferences(SolicitorReferences solicitorReferences) {
-        return SolicitorReferences
-            .builder()
-            .applicantSolicitor1Reference(
-                ofNullable(solicitorReferences)
-                    .map(SolicitorReferences::getApplicantSolicitor1Reference)
-                    .orElse("Not Provided"))
-            .respondentSolicitor1Reference(
-                ofNullable(solicitorReferences)
-                    .map(SolicitorReferences::getRespondentSolicitor1Reference)
-                    .orElse("Not Provided"))
             .build();
     }
